feat(book): implement get all books to index and delete book with valid isbn

+++ 14_library-management-system_practice/Controllers/BookController.cs
@@ -10,16 +10,27 @@ namespace _14_library_management_system_practice.Controllers
 
 
         [HttpGet]
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
-            var books = _repo.getAllBooksAsync();
+            var books = await _repo.getAllBooksAsync();
             return View(books);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, string? isbn)
         {
+            var book = await _repo.GetBookByIdAsync(id);
+            if (book is null) return NotFound();
+
+            if (string.IsNullOrWhiteSpace(isbn) || !isbn.Equals(book.ISBN))
+            {
+                TempData["Error"] = "ISBN không đúng. Không thể xoá!";
+                return RedirectToAction(nameof(Index));
+            }
+
             await _repo.DeleteBookAsync(id);
+            TempData["Success"] = "Đã xoá sách thành công.";
+
             return RedirectToAction(nameof(Index));
         }
     }
